Move Form folder into mainfolder. Change js to jsx&
